Continuous lift positions + chairlift redesign + PBR terrain
- Building.Pos and Lift.Base/Top move from cell indices to mgl32.Vec2;
  PlaceLift/PlaceBuilding take continuous coords throughout sim and
  scenario code.
- tower.scad authored to match the 18 m cable height; lift_station
  rescaled for ground-level boarding with a longer beam pushing the
  bullwheel out from the column.
- Procedural T-tower mesh dropped in favour of an OBJ-instanced
  static batch; end towers are skipped so they don't sit atop the
  station bullwheel beams.
- Cable Y now ramps from BullwheelHeight at each station up to
  TowerHeight via CableHeightAt(frac, length); chairs follow the same
  profile.
- Terrain shader switched to triplanar PBR using Poly Haven CC0
  snow_02 + rock_face_03 (diffuse/normal/roughness), with anti-tiling
  multi-scale snow blend and a slope-driven rock/snow split decoupled
  from the triplanar weights.

+++ internal/save/format.go
@@ -30,11 +30,12 @@ type ObjectData struct {
 }
 
 // BuildingData is a placed lodge. ID is preserved across save/load so that
-// agent.TargetID references stay valid.
+// agent.TargetID references stay valid. X/Z are continuous world XZ
+// (metres); Y is reconstructed from terrain at load time.
 type BuildingData struct {
 	ID            uint64  `json:"id,omitempty"`
-	X             int     `json:"x"`
-	Z             int     `json:"z"`
+	X             float32 `json:"x"`
+	Z             float32 `json:"z"`
 	Rotation      float32 `json:"r,omitempty"`
 	MeanSpawnRate float64 `json:"mean_spawn_rate"`
 	SkierCount    int     `json:"skier_count,omitempty"`
@@ -49,13 +50,14 @@ type ChairData struct {
 
 // LiftData is a placed lift, including its full runtime state (chair
 // positions and passenger references, queue order) so a save round-trips
-// without freezing skiers in mid-air.
+// without freezing skiers in mid-air. Base/Top are continuous world XZ
+// (metres).
 type LiftData struct {
 	ID       uint64      `json:"id,omitempty"`
-	BaseX    int         `json:"bx"`
-	BaseZ    int         `json:"bz"`
-	TopX     int         `json:"tx"`
-	TopZ     int         `json:"tz"`
+	BaseX    float32     `json:"bx"`
+	BaseZ    float32     `json:"bz"`
+	TopX     float32     `json:"tx"`
+	TopZ     float32     `json:"tz"`
 	Speed    float32     `json:"speed,omitempty"`
 	Chairs   []ChairData `json:"chairs,omitempty"`
 	QueueIDs []uint64    `json:"queue,omitempty"`
